맵생성시 카메라가 이동안하는 문제
씨네머신 오브젝트를 비활성화, 활성화 하니까 원하는 상황이 작동이 됨

+++ 2D_Project/Assets/Scripts/PlayerStats.cs
@@ -72,7 +72,7 @@ public class PlayerStats : MonoBehaviour
         GameManager.instance.SavePlayerStats(this);
     }
 
-    public void UpgradeAttackSpeed(int amount)
+    public void UpgradeAttackSpeed(float amount)
     {
         attackSpeed += amount;
         GameManager.instance.SavePlayerStats(this);
@@ -84,7 +84,7 @@ public class PlayerStats : MonoBehaviour
         GameManager.instance.SavePlayerStats(this);
     }
 
-    public void UpgradeMoveSpeed(int amount)
+    public void UpgradeMoveSpeed(float amount)
     {
         moveSpeed += amount;
         GameManager.instance.SavePlayerStats(this);
